base

+++ src/base/FormatUtils.h
@@ -19,7 +19,7 @@ NB_CORE_EXPORT UnicodeString Sprintf(const UnicodeString fmt, fmt::ArgList args)
 FMT_VARIADIC_W(const UnicodeString, Sprintf, const UnicodeString)
 
 NB_CORE_EXPORT UnicodeString FmtLoadStr(int32_t id, fmt::ArgList args);
-FMT_VARIADIC_W(const UnicodeString, FmtLoadStr, intptr_t)
+FMT_VARIADIC_W(const UnicodeString, FmtLoadStr, int32_t)
 
 } // namespace nb
